requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0

# RAG 相关依赖
chromadb>=0.5.0
openai==1.10.0
pypdf==4.0.1
python-multipart==0.0.6
langchain-text-splitters==0.0.1
numpy==1.26.4

# 数据库（PostgreSQL）
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1